pyproject.toml
[project]
name = "django-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2rc1",
    "django-filter>=25.1",
    "djangorestframework>=3.15.2",
    "gunicorn>=23.0.0",
    "markdown>=3.7",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "coverage>=7.7.1",
    "ipython>=9.0.2",
]
